+++ lib/constants.dart
@@ -1,9 +1,3 @@
 import 'package:flutter/material.dart';
 
 const Color primaryColour = Color(0xff545ad8);
-const TextStyle dashboardTextStyle = TextStyle(
-  color: Colors.black,
-  fontSize: 12,
-  fontFamily: 'Montserrat',
-  fontWeight: FontWeight.w500,
-);
