+++ pubspec.yaml
@@ -62,8 +62,6 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
     - images/
-    - images/bottomnavigationbar/
-    - images/dashboard/
   #   - images/a_dot_ham.jpeg
 
   # An image asset can refer to one or more resolution-specific "variants", see
